+++ headerparser/scanner.py
@@ -1,6 +1,7 @@
 import re
-from   .errors import MalformedHeaderError, UnexpectedFoldingError
-from   .util   import ascii_splitlines
+from   warnings import warn
+from   .errors  import MalformedHeaderError, UnexpectedFoldingError
+from   .util    import ascii_splitlines
 
 def scan_string(s, **kwargs):
     """
@@ -8,11 +9,10 @@ def scan_string(s, **kwargs):
     ``(name, value)`` pairs for each header field in the input, plus a ``(None,
     body)`` pair representing the body (if any) after the header section.
 
-    See `scan_lines()` for more information on the exact behavior of the
-    scanner.
+    See `scan()` for more information on the exact behavior of the scanner.
 
     :param s: a string which will be broken into lines on CR, LF, and CR LF
-        boundaries and passed to `scan_lines()`
+        boundaries and passed to `scan()`
     :param kwargs: :ref:`scanner options <scan_opts>`
     :rtype: generator of pairs of strings
     :raises MalformedHeaderError: if an invalid header line, i.e., a line
@@ -20,7 +20,7 @@ def scan_string(s, **kwargs):
     :raises UnexpectedFoldingError: if a folded (indented) line that is not
         preceded by a valid header line is encountered
     """
-    return scan_lines(ascii_splitlines(s), **kwargs)
+    return scan(ascii_splitlines(s), **kwargs)
 
 def scan_file(fp, **kwargs):
     """
@@ -28,11 +28,13 @@ def scan_file(fp, **kwargs):
     ``(name, value)`` pairs for each header field in the input, plus a ``(None,
     body)`` pair representing the body (if any) after the header section.
 
-    See `scan_lines()` for more information on the exact behavior of the
-    scanner.
+    See `scan()` for more information on the exact behavior of the scanner.
+
+    .. deprecated:: 0.4.0
+        Use `scan()` instead.
 
     :param fp: A file-like object than can be iterated over to produce lines to
-        pass to `scan_lines()`.  Opening the file in universal newlines mode is
+        pass to `scan()`.  Opening the file in universal newlines mode is
         recommended.
     :param kwargs: :ref:`scanner options <scan_opts>`
     :rtype: generator of pairs of strings
@@ -41,19 +43,44 @@ def scan_file(fp, **kwargs):
     :raises UnexpectedFoldingError: if a folded (indented) line that is not
         preceded by a valid header line is encountered
     """
-    ### TODO: Handle files not opened in universal newlines mode?
-    return scan_lines(fp, **kwargs)
+    warn('scan_file() is deprecated.  Use scan() instead.', DeprecationWarning)
+    return scan(fp, **kwargs)
 
-def scan_lines(
-    iterable,
-    separator_regex       = re.compile(r'[ \t]*:[ \t]*'),
-    skip_leading_newlines = False,
-):
+def scan_lines(fp, **kwargs):
     """
     Scan an iterable of lines for RFC 822-style header fields and return a
     generator of ``(name, value)`` pairs for each header field in the input,
     plus a ``(None, body)`` pair representing the body (if any) after the
     header section.
+
+    See `scan()` for more information on the exact behavior of the scanner.
+
+    .. deprecated:: 0.4.0
+        Use `scan()` instead.
+
+    :param iterable: an iterable of strings representing lines of input
+    :param kwargs: :ref:`scanner options <scan_opts>`
+    :rtype: generator of pairs of strings
+    :raises MalformedHeaderError: if an invalid header line, i.e., a line
+        without either a colon or leading whitespace, is encountered
+    :raises UnexpectedFoldingError: if a folded (indented) line that is not
+        preceded by a valid header line is encountered
+    """
+    warn('scan_lines() is deprecated.  Use scan() instead.', DeprecationWarning)
+    return scan(fp, **kwargs)
+
+def scan(
+    iterable,
+    separator_regex       = re.compile(r'[ \t]*:[ \t]*'),
+    skip_leading_newlines = False,
+):
+    """
+    .. versionadded:: 0.4.0
+
+    Scan a text-file-like object or iterable of lines for RFC 822-style header
+    fields and return a generator of ``(name, value)`` pairs for each header
+    field in the input, plus a ``(None, body)`` pair representing the body (if
+    any) after the header section.
 
     Each field value is a single string, the concatenation of one or more
     lines, with leading whitespace on lines after the first preserved.  The
@@ -77,7 +104,8 @@ def scan_lines(
     option is `False` (the default), then all other lines will be treated as
     part of the body and will not be scanned for header fields.
 
-    :param iterable: an iterable of strings representing lines of input
+    :param iterable: a text-file-like object or iterable of strings
+        representing lines of input
     :param kwargs: :ref:`scanner options <scan_opts>`
     :rtype: generator of pairs of strings
     :raises MalformedHeaderError: if an invalid header line, i.e., a line