+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 v0.4.0 (in development)
 -----------------------
+- Added a `scan()` function combining the behavior of `scan_file()` and
+  `scan_lines()`, which are now deprecated
 
 v0.3.0 (2018-10-12)
 -------------------